usb: f_gsi: Prevent enabling data path if disconnect follows connect
In the state CONNECT_IN_PROGRESS if EVT_CONNECT is
followed by a EVT_DISCONNECT event, there is a possible case
where doorbell is unblocked with endpoints disabled resulting
in a possible spurious doorbell from the USB wrapper.

Fix this by bailing out if a connect event is followed
by disconnect.

Change-Id: I73f971548b6c13924e9f857e57fa6dd517a4e56a
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>